Master_Thesis_Paulogcd_2025.HP

Master_Thesis_Paulogcd_2025.Health_Proxy_plot

Master_Thesis_Paulogcd_2025.average_health_plot

Master_Thesis_Paulogcd_2025.health

Master_Thesis_Paulogcd_2025.health_transition_3D_plot

Master_Thesis_Paulogcd_2025.plot_demographic_comparison

Master_Thesis_Paulogcd_2025.plot_population_weather

Master_Thesis_Paulogcd_2025.population_health_simulation

Master_Thesis_Paulogcd_2025.produce_graph

Master_Thesis_Paulogcd_2025.survival

Master_Thesis_Paulogcd_2025.temperature

Master_Thesis_Paulogcd_2025.test

Master_Thesis_Paulogcd_2025.HP — Method

The HP function is not to be used alone. It returns the estimated Health Proxy value for a combination of age and temperature values.

Master_Thesis_Paulogcd_2025.Health_Proxy_plot — Method

The Health_Proxy_plot function plots the predicted health proxy for a range of age, and for the different temperature scenarios.

Master_Thesis_Paulogcd_2025.average_health_plot — Method

The average_health_plot function generates a plot of the average health over time for the three different temperature scenarios. It takes N and T as arguments, with N the size of the population, and T the number of periods.

Master_Thesis_Paulogcd_2025.health — Method

The health function is not to be used alone. It returns the estimated health transition probabilities for a combination of age, temperature, and previous health state values.

Master_Thesis_Paulogcd_2025.health_transition_3D_plot — Method

The health_transition_3D_plot function generates 3D plots of the health transition probabilities. It produces 5 html files, which correspond to transition probabilities from the 5 possible health states.

Master_Thesis_Paulogcd_2025.plot_demographic_comparison — Method

This function plots the demographic path of the four defined trajectories.

Master_Thesis_Paulogcd_2025.plot_population_weather — Method

This function is not to be used alone.

Master_Thesis_Paulogcd_2025.population_health_simulation — Method

The population_health_simulation function is not to be used alone. It produces a population simulation in which individuals do not draw their living status, but draw their health status each period.

Master_Thesis_Paulogcd_2025.produce_graph — Method

This function is not to be used alone. The produce_graph function generates a graph with three different populations with three different temperature paths.

Master_Thesis_Paulogcd_2025.survival — Method

This function is not to be used alone. The survival function computes the probability of survival for a combination of age and temperature.

Master_Thesis_Paulogcd_2025.temperature — Method

The temperature function returns the plot of average annual temperature from 1900 to 2022.

Master_Thesis_Paulogcd_2025.test — Method

The test function ensures the package is well loaded.